requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
plotly
streamlit_audio_recorder
openpyxl
